Put Mapping section in box instead of accordion.
